Restructured to accomodate David's design plans
There is a package.json in the app folder, it does nothing besides
allow electron to run the app folder. The package.json in the top
level directory is for devdependencies only and allows one to run
makefile scripts OR npm scripts to run the program.

Currently only makefile scripts package up the program.