Properly display wallet addresses and transactions
I think the changes should speak for themselves but essentially the
two tables in the Wallet plugin should be populated based on a user's
transaction history.

What's not considered are addresses that the user hasn't had a
transaction be related with yet. Those are just forgotten and can't be
retrieved by API call